Use tag instead of direct name for updatevm setting
Using direct name (whonix-gw/whonix-ws) to set updates proxy VM isn't
the best idea, because it won't work for template clones. This is
especially bad for whonix, because it results in clearnet access to VMs
that should be behind tor (downloading updates over clearnet in this
case).

Fixes QubesOS/qubes-issues#3118