Exclude controllers listed via rd.qubes.dom0_usb option from sys-usb
When multiple USB controllers are available, one of the possibilities is
to keep some in dom0 (for HID) and attach the rest to sys-usb. Initramfs
will use rd.qubes.dom0_usb option to select which remain in dom0, and
let the initial setup follow the same logic when creating sys-usb.

QubesOS/qubes-issues#2116