Fix policy for keyboard/mouse with sys-gui-gpu
When the action is 'ask default_target=sys-gui-gpu', there needs to be
another rule actually allowing the sys-gui-gpu target. Contrary to the
'allow target=sys-gui-gpu', the ask's default_target= parameter doesn't
implicitly allow that target.

Add the rule for allow (and deny) too for simplicy, but it's actually
needed only for the 'ask' action.

Fixes QubesOS/qubes-issues#9980